Move applyQuery flat-SQL tests into StreamingTest, delete standalone file
