+++ spark/src/test/scala/ai/chronon/spark/test/StreamingTest.scala
@@ -21,6 +21,7 @@ import ai.chronon.api
 import ai.chronon.api.{Accuracy, Builders, Constants, Operation, TimeUnit, Window}
 import ai.chronon.api.Constants.ChrononMetadataKey
 import ai.chronon.api.Extensions._
+import ai.chronon.api.QueryUtils
 import ai.chronon.spark.test.StreamingTest.buildInMemoryKvStore
 import ai.chronon.online.MetadataStore
 import ai.chronon.spark.Extensions._
@@ -112,5 +113,86 @@ class StreamingTest extends TestCase {
     metadataStore.putJoinConf(joinConf)
     joinConf.joinParts.toScala.foreach(jp =>
       OnlineUtils.serve(tableUtils, inMemoryKvStore, buildInMemoryKvStore, namespace, today, jp.groupBy))
+  }
+
+  // ---------------------------------------------------------------------------
+  // applyQuery flat-SQL semantics tests
+  //
+  // These verify the invariant that JoinSourceRunner.applyQuery enforces after
+  // replacing the two-stage selectExpr→filter with temp-view + session.sql:
+  // wheres resolve against the pre-select (raw) input schema, matching offline.
+  // ---------------------------------------------------------------------------
+
+  private def makeRawView(viewName: String): Unit = {
+    import spark.implicits._
+    Seq(
+      (1L, "alice", "US"),
+      (0L, "bob",   "US"),
+      (2L, "carol", "CA")
+    ).toDF("raw_user_id", "raw_name", "raw_country")
+      .createOrReplaceTempView(viewName)
+  }
+
+  def testWheresOnRawColumnBeforeRename(): Unit = {
+    makeRawView("test_raw_column")
+    val sql = QueryUtils.build(
+      selects = Map("user_id" -> "raw_user_id", "name" -> "raw_name"),
+      from = "test_raw_column",
+      wheres = Seq("raw_user_id > 0")
+    )
+    val result = spark.sql(sql).collect()
+    assert(result.length == 2, s"Expected 2 rows, got ${result.length}")
+    val userIds = result.map(_.getAs[Long]("user_id")).toSet
+    assert(userIds == Set(1L, 2L), s"Expected {1, 2}, got $userIds")
+  }
+
+  def testWheresOnColumnAbsentFromSelects(): Unit = {
+    makeRawView("test_absent_from_selects")
+    val sql = QueryUtils.build(
+      selects = Map("user_id" -> "raw_user_id", "name" -> "raw_name"),
+      from = "test_absent_from_selects",
+      wheres = Seq("raw_country = 'US'")
+    )
+    val result = spark.sql(sql).collect()
+    assert(result.length == 2, s"Expected 2 rows (US only), got ${result.length}")
+    val names = result.map(_.getAs[String]("name")).toSet
+    assert(names == Set("alice", "bob"), s"Expected {alice, bob}, got $names")
+  }
+
+  def testWheresOnSelectAliasFails(): Unit = {
+    makeRawView("test_alias_in_where")
+    val sql = QueryUtils.build(
+      selects = Map("user_id" -> "raw_user_id", "name" -> "raw_name"),
+      from = "test_alias_in_where",
+      wheres = Seq("user_id > 0")
+    )
+    try {
+      spark.sql(sql).collect()
+      throw new AssertionError("Expected AnalysisException for alias reference in WHERE, but query succeeded")
+    } catch {
+      case _: org.apache.spark.sql.AnalysisException => // expected: WHERE evaluates before SELECT
+    }
+  }
+
+  def testApplyQueryNoSelects(): Unit = {
+    makeRawView("test_no_selects")
+    val sql = QueryUtils.build(
+      selects = null,
+      from = "test_no_selects",
+      wheres = Seq("raw_user_id > 0")
+    )
+    val result = spark.sql(sql).collect()
+    assert(result.length == 2, s"Expected 2 rows with SELECT *, got ${result.length}")
+  }
+
+  def testApplyQueryNoWheres(): Unit = {
+    makeRawView("test_no_wheres")
+    val sql = QueryUtils.build(
+      selects = Map("user_id" -> "raw_user_id"),
+      from = "test_no_wheres",
+      wheres = Seq.empty
+    )
+    val result = spark.sql(sql).collect()
+    assert(result.length == 3, s"Expected all 3 rows, got ${result.length}")
   }
 }
